[+] очень много запросов, на винил, на машинки, на бензин крч много всего сделал, нет сил описывать

+++ streetrace-api/src/main/java/com/streetrace/api/dto/car/UserCarVinylDTO.java
@@ -11,7 +11,7 @@ import lombok.*;
 public class UserCarVinylDTO {
     private Long id;
     private String name;
-    private int x; // Координата X
-    private int y; // Координата Y
+    private float x; // Координата X
+    private float y; // Координата Y
     private float rotationAngle; // Угол поворота винила
 }
